+++ open-reac/src/main/java/com/powsybl/openreac/exceptions/InvalidParametersException.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.openreac.parameters;
+package com.powsybl.openreac.exceptions;
 
 import com.powsybl.commons.PowsyblException;
